backend/requirements.txt
tensorflow==2.15.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.3
